pyproject.toml
[tool.poetry]
name = "song_translator"
version = "0.1.0"
description = ""
authors = ["Anatoli"]

[tool.poetry.dependencies]
python = "^3.8"
google-trans-new = "1.1.9"
Django = "3.2"
djangorestframework = "3.12.4"
requests = "2.25.1"
psycopg2 = "2.8.6"
django-cors-headers = "^3.7.0"
django-model-utils = "^4.1.1"
django-query-logger = "^0.1.2"
django-factory-boy-generator = "^1.0.3"
django-filter = "^2.4.0"
django-factory-generator = "^1.0.4"
launchdarkly-server-sdk = "^7.1.0"
python-dotenv = "^0.17.1"
django-environ = "^0.4.5"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
